build: Disable compiler warning C4251
The protobuf support library uses templates in its public interfaces, and
MSVC complains about the template classes not being exported from the built
DLL. As far as we can tell, the warning is wrong and not actionable. It's
not actionable because, according to the protobuf developers, there is "no
reasonable way" to export template classes[1]. It's wrong because (at least
in protobuf's template usage) complete template definitions are available in
header files, so all client code ends up compiling its own instances.

1. https://github.com/protocolbuffers/protobuf/blob/main/cmake/README.md#notes-on-compiler-warnings